Copy List with Random pointer ( Brain Teaser )
Link: https://leetcode.com/problems/copy-list-with-random-pointer/
This question is one my favorite question. This is a perfect blend of Concept , Implementation and yes of course require an Idea Click.
Basically copy of linked list can be done easily but to copy the random pointer is quite difficult Task.
Approach:
Create a new linked list with copied value from the original one.
Then for Random pointers , Bypassing ( In my terminology ) is required.  
That is we somehow connect a way from start node ( for which random pointer to set ) of cloned list to the original list and then get its
random pointer value and then bypass to the cloned list i.e destination node.
To construct the way : 
1 . Set the Random pointer of cloned list to point to the node of original list having same data value
2 . Then , set the next pointer of the original list to point to the node of cloned list having same data value.
( Try to find out why an extra storage (map) is required )
set the cloned random pointer as : 
                 clone->random = clone->random->random->next